+++ docs/history.md
@@ -5,9 +5,9 @@
 
 ### _v1.19.0_ - September 18th, 2023
 
-- Changed the following babies:
-  - Cupid Baby (#289) - Base Larynx effect every 8 tears. (Thanks Gamonymous)
-  - Locust Baby (#345) - Starts with 10 special Abyss locusts + blindfolded. (Thanks Gamonymous)
+- The following babies have new effects:
+  - Cupid Baby (#289) - "Starts with Cupid's Arrow" --> "Base Larynx effect every 8 tears" (Thanks Gamonymous)
+  - Locust Baby (#345) - "Starts with 20 Abyss locusts + blindfolded" --> "Starts with 10 special Abyss locusts + blindfolded"
   - Multidimensional Baby (#593) - "Enemies spawn 3 Swarm Spiders on death" --> "Enemies spawn 2 Swarm Spiders on death"
 
 ### _v1.18.0_ - September 17th, 2023