+++ docs/history.md
@@ -9,6 +9,7 @@
   - Cupid Baby (#289) - "Starts with Cupid's Arrow" --> "Base Larynx effect every 8 tears"
   - Locust Baby (#345) - "Starts with 20 Abyss locusts + blindfolded" --> "Starts with 10 special Abyss locusts + blindfolded"
   - Multidimensional Baby (#593) - "Enemies spawn 3 Swarm Spiders on death" --> "Enemies spawn 2 Swarm Spiders on death"
+- Fixed the bug where having two or more Brimstones would cause cause an error. (Thanks mgln)
 
 ### _v1.18.0_ - September 17th, 2023